Replace the custom exception thrown by integer return code to be Pure C Interface.

+++ src/base64.h
@@ -8,7 +8,7 @@ namespace jed_utils
     class Base64
     {
     public:
-    	static std::string Encode(unsigned char const *bytes_to_encode, unsigned int in_len);
+    	static std::string Encode(unsigned char const *bytes_to_encode, size_t in_len);
 	    static std::string Decode(std::string const &encoded_string);
     };
 } // namespace jed_utils
